+++ beekeeper-path-cleanup/src/test/java/com/expediagroup/beekeeper/path/cleanup/service/PagingCleanupServiceTest.java
@@ -213,7 +213,6 @@ public class PagingCleanupServiceTest {
         .housekeepingStatus(housekeepingStatus)
         .creationTimestamp(localNow)
         .modifiedTimestamp(localNow)
-        .modifiedTimestamp(localNow)
         .cleanupDelay(PeriodDuration.of(Duration.parse("P3D")))
         .cleanupAttempts(0)
         .lifecycleType(UNREFERENCED.toString())