+++ beekeeper-scheduler-apiary/src/main/java/com/expediagroup/beekeeper/scheduler/apiary/context/CommonBeans.java
@@ -41,6 +41,7 @@ import com.expediagroup.beekeeper.core.model.LifecycleEventType;
 import com.expediagroup.beekeeper.core.repository.BeekeeperHistoryRepository;
 import com.expediagroup.beekeeper.core.service.BeekeeperHistoryService;
 import com.expediagroup.beekeeper.scheduler.apiary.filter.EventTypeListenerEventFilter;
+import com.expediagroup.beekeeper.scheduler.apiary.filter.IcebergTableListenerEventFilter;
 import com.expediagroup.beekeeper.scheduler.apiary.filter.ListenerEventFilter;
 import com.expediagroup.beekeeper.scheduler.apiary.filter.LocationOnlyUpdateListenerEventFilter;
 import com.expediagroup.beekeeper.scheduler.apiary.filter.TableParameterListenerEventFilter;
@@ -98,7 +99,8 @@ public class CommonBeans {
         new EventTypeListenerEventFilter(eventClasses),
         new LocationOnlyUpdateListenerEventFilter(),
         new TableParameterListenerEventFilter(),
-        new WhitelistedListenerEventFilter()
+        new WhitelistedListenerEventFilter(),
+        new IcebergTableListenerEventFilter()
     );
 
     return new MessageEventHandler(generator, filters);
@@ -122,7 +124,8 @@ public class CommonBeans {
 
     List<ListenerEventFilter> filters = List.of(
         new EventTypeListenerEventFilter(eventClasses),
-        new TableParameterListenerEventFilter()
+        new TableParameterListenerEventFilter(),
+        new IcebergTableListenerEventFilter()
     );
 
     return new MessageEventHandler(generator, filters);