+++ beekeeper-scheduler-apiary/src/test/java/com/expediagroup/beekeeper/scheduler/apiary/generator/UnreferencedHousekeepingPathGeneratorTest.java
@@ -18,7 +18,7 @@ package com.expediagroup.beekeeper.scheduler.apiary.generator;
 import static java.lang.String.format;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.junit.Assert.fail;
+import static org.junit.jupiter.api.Assertions.fail;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;